Refactor environment into package and simplify
