+++ pyproject.toml
@@ -9,5 +9,6 @@ description = "Gym compatible Tetris environment"
 readme = "README.md"
 requires-python = ">=3.10"
 
-[tool.setuptools]
-py-modules = ["tetris_env"]
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["tetris_env*"]
